feat: improve hashing and add scrypt support

+++ extract.sh
@@ -27,22 +27,38 @@ else
   password="$PASSWORD"
 fi
 
+password=$(echo -n "$password")
 echo "Generate hashes..."
 
 zip_pwd=$(echo -n "$password" | sha512sum | cut -f 1 -d " ")
-zip_pwd=$(echo "$zip_pwd" $(echo -n "$password") "$zip_pwd" | sha512sum | cut -f 1 -d " ")
-gpg_pwd=$(echo "$zip_pwd" $(echo -n "$password") "$zip_pwd"  | sha512sum | cut -f 1 -d " ")
+zip_pwd=$(echo "$zip_pwd" "$password" "$zip_pwd" | sha512sum | cut -f 1 -d " ")
+gpg_pwd=$(echo "$zip_pwd" "$password" "$zip_pwd" "$password"  | sha512sum | cut -f 1 -d " ")
+scr_pwd=$(echo "$zip_pwd" "$password" "$gpg_pwd")
 
 for i in {1..1000}; do
-  zip_pwd=$(echo "$zip_pwd" "$gpg_pwd" | sha512sum | cut -f 1 -d " ")
-  gpg_pwd=$(echo "$gpg_pwd" "$zip_pwd" | sha512sum | cut -f 1 -d " ")
+  zip_pwd=$(echo "$zip_pwd" "$password" "$gpg_pwd" "$scr_pwd" | sha512sum | cut -f 1 -d " ")
+  gpg_pwd=$(echo "$scr_pwd" "$gpg_pwd" "$password" "$zip_pwd" | sha512sum | cut -f 1 -d " ")
+  if [ "$3" = "-s" ]; then
+    scr_pwd=$(echo "$scr_pwd" "$zip_pwd" "$password" "$gpg_pwd" | sha512sum | cut -f 1 -d " ")
+  fi
 done
 
 echo "Decrypting to '.tmp.zip' ..."
 
-gpg --batch --passphrase "$gpg_pwd" -d "$1" > .tmp.zip
-echo "Extracting..."
-unzip -P "$zip_pwd" -d "$2" -o ".tmp.zip"
-echo "Remove '.tmp.zip' ..."
-rm ".tmp.zip"
-echo "Done. Extracted path: " "$2"
+if [ "$3" = "-s" ]; then
+  export scrypt_pwd=$scr_pwd
+  scrypt dec --passphrase "env:scrypt_pwd" "$1" | gpg --batch --passphrase "$gpg_pwd" -d > .tmp.zip
+  unset scrypt_pwd
+else
+  gpg --batch --passphrase "$gpg_pwd" -d "$1" > .tmp.zip
+fi
+
+if [ -f ".tmp.zip" ]; then
+  echo "Extracting..."
+  unzip -P "$zip_pwd" -d "$2" -o ".tmp.zip"
+  echo "Remove '.tmp.zip' ..."
+  rm ".tmp.zip"
+  echo "Done. Extracted path: " "$2"
+else
+  echo "Something went wrong. Check that your password is correct."
+fi
